Metadata and Path fixes
Fixed the path of books in Library. Fixed Metadata in factions. Split Offtime form into client and server, then added Metadata.
