+++ src/app/factions/page.tsx
@@ -4,7 +4,7 @@ import { Metadata } from "next";
 export const metadata: Metadata = {
   title: "Factions",
   description:
-    "Explore the pantheon of gods and goddesses that shape the destiny of Teothe, with domains, hierarchy, and legendary tales awaiting discovery.",
+    "Discover the influential factions of Teothe, their goals, alliances, and rivalries that shape Teothe’s politics, conflicts, and history.",
 };
 
 export default function FactionsPage() {
